requirements.txt
anyio==4.4.0
asyncapi==0.14.1
beautifulsoup4==4.12.3
broadcaster==0.3.1
certifi==2024.7.4
charset-normalizer==3.3.2
dictdaora==0.0.2
exceptiongroup==1.2.2
fake-useragent-ex==0.1.12
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
jsondaora==0.14.1
orjson==3.10.7
packaging==24.1
playwright==1.46.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyee==11.1.0
pytest==8.3.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
text-unidecode==1.3
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
